refactor(test): test 패키지 구조 전반 수정
- `E2ETest/...` -> `test/e2e/...` 로 이관
- `UnitTest/...` -> `test/integration/...`로 이관(jpa 관련 통합 테스트에 가깝기 때문에)
