+++ src/test/java/gift/Test/integration/jpa/AbstractRepositoryTest.java
@@ -1,0 +1,12 @@
+package gift.Test.integration.jpa;
+
+import org.junit.jupiter.api.MethodOrderer;
+import org.junit.jupiter.api.TestMethodOrder;
+import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
+import org.springframework.test.context.ActiveProfiles;
+
+@ActiveProfiles("test")
+@TestMethodOrder(MethodOrderer.OrderAnnotation.class)
+@DataJpaTest
+public abstract class AbstractRepositoryTest {
+}
