--- src/test/java/gift/Test/repository/AbstractRepositoryTest.java
@@ -1,12 +1,0 @@
-package gift.Test.repository;
-
-import org.junit.jupiter.api.MethodOrderer;
-import org.junit.jupiter.api.TestMethodOrder;
-import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
-import org.springframework.test.context.ActiveProfiles;
-
-@ActiveProfiles("test")
-@TestMethodOrder(MethodOrderer.OrderAnnotation.class)
-@DataJpaTest
-public abstract class AbstractRepositoryTest {
-}
